% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernadette-package.R
\docType{package}
\name{Bernadette-package}
\alias{Bernadette-package}
\alias{Bernadette}
\title{Bayesian inference and model selection for stochastic epidemics}
\description{
\code{Bernadette} provides Bayesian analysis for stochastic extensions of dynamic non-linear systems using advanced computational algorithms.
}
\references{
Bouranis, L., Demiris, N. Kalogeropoulos, K. and Ntzoufras, I. (2022). Bayesian analysis of diffusion-driven multi-type epidemic models with application to COVID-19. arXiv: \url{https://arxiv.org/abs/2211.15229}

Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.3. \url{https://mc-stan.org}
}
