% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_matrices}
\alias{contact_matrices}
\title{Contact matrices per country}
\format{
A list of 16 by 16 dataframes for 177 countries.
}
\usage{
data(contact_matrices)
}
\value{
A list object of 16 by 16 dataframes for 177 countries.
}
\description{
A list of 16 by 16 contact matrices for 177 countries.
Row i of a column j of a contact matrix corresponds to the number of contacts
made by an individual in group i with an individual in group j.
}
\section{References}{

Prem, K., van Zandvoort, K., Klepac, P. et al (2020). Projecting contact matrices in 177 geographical regions: an update and comparison with empirical data for the COVID-19 era. medRxiv 2020.07.22.20159772; doi: https://doi.org/10.1101/2020.07.22.20159772
}

\keyword{datasets}
