% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{contact_matrix}
\alias{contact_matrix}
\title{Country-specific contact matrix}
\usage{
contact_matrix(country)
}
\arguments{
\item{country}{A character indicating the country identifier. See \link[Bernadette]{country_contact_matrices}.}
}
\value{
An object of class "data.frame".
}
\description{
A 16 by 16 contact matrix whose row i of a column j corresponds to the number of contacts
made by an individual in group i with an individual in group j.
}
\examples{
conmat <- contact_matrix(country = "GRC")

}
\references{
Prem, K., van Zandvoort, K., Klepac, P. et al (2020). Projecting contact matrices in 177 geographical regions: an update and comparison with empirical data for the COVID-19 era. medRxiv 2020.07.22.20159772; doi: https://doi.org/10.1101/2020.07.22.20159772
}
