% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_un.R
\name{countries_un}
\alias{countries_un}
\title{Names of countries with an available age distribution}
\usage{
countries_un()
}
\description{
Function to extract the names of the countries whose discrete age distribution is
available by the United Nations.
}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2019). World Population Prospects 2019, Online Edition. Rev. 1.
}
