% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_contact_matrix.R
\name{country_contact_matrices}
\alias{country_contact_matrices}
\title{Names of countries for which a contact matrix is available}
\usage{
country_contact_matrices()
}
\description{
Function to extract the names of the countries whose projected contact matrix is available,
}
\references{
Prem, K., van Zandvoort, K., Klepac, P. et al (2017). Projecting contact matrices in 177 geographical regions: an update and comparison with empirical data for the COVID-19 era. medRxiv 2020.07.22.20159772; doi: https://doi.org/10.1101/2020.07.22.20159772
}
