% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_population.R
\name{country_population}
\alias{country_population}
\title{Country-specific total population}
\usage{
country_population(country, year = NULL)
}
\arguments{
\item{country}{character; country identifier, following the List of United Nations Member States. See \link[Bernadette]{countries_un}.}

\item{year}{numeric; calendar year.}
}
\value{
An object of type "numeric", containing the population figure for a given country and calendar year.
}
\description{
Function to extract the total population of a country for a given year, based on
the United Nations 2019 Revision of World Population Prospects.
}
\examples{

\dontrun{
country_pop <- country_population(country = "Greece",
                                 year    = 2020)
}

}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2019). World Population Prospects 2019, Online Edition. Rev. 1.
}
