% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographics_un}
\alias{demographics_un}
\title{Age distribution per country}
\format{
A data frame with 1,164,891 rows and 7 variables:
\describe{
  \item{Location}{character; name of country}
  \item{Time}{numeric; year}
  \item{AgeGrp}{character; 5-year age bands}
  \item{AgeGrpStart}{numeric; year indicating the start of the respective age group}
  \item{PopMale}{numeric; male population in the given age group}
  \item{PopFemale}{numeric; female population in the given age group}
  \item{PopTotal}{total population in the given age group}
}
}
\usage{
data(demographics_un)
}
\description{
A dataset containing the age distribution per country and per year. Information is broken
down by gender and 5-year age bands, following the United Nations 2019 Revision
of World Population Prospects.
}
\keyword{datasets}
