% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pop_age_gender.R
\name{extract_pop_age_gender}
\alias{extract_pop_age_gender}
\title{Country-specific population by gender and age group, with 5-year bands.}
\usage{
extract_pop_age_gender(country, year, stratify_gender = FALSE)
}
\arguments{
\item{country}{character; country identifier, following the \href{https://www.un.org/en/about-us/member-states}{List of United Nations Member States}.}

\item{year}{numeric; calendar year}

\item{stratify_gender}{logical; TRUE/FALSE: aggregate the population by gender only.}
}
\value{
A data frame that contains the portion of the population per grouping.
}
\description{
Function to extract the population of a country for a given year, broken
down by gender and 5-year age bands, following the United Nations 2019 Revision
of World Population Prospects.
}
\examples{

\dontrun{
check_age <- extract_pop_age_gender(country         = "Greece",
                                   year            = 2020,
                                   stratify_gender = FALSE)
}

}
