% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cases_deaths.R
\name{import_cases_deaths}
\alias{import_cases_deaths}
\title{Country-specific COVID-19 daily time series summary table}
\usage{
import_cases_deaths(country, start_date = NULL, end_date = NULL)
}
\arguments{
\item{country}{character; country identifier, following the \href{https://www.un.org/en/about-us/member-states}{List of United Nations Member States}.}

\item{start_date}{character; calendar date in the format "yyyy-mm-dd"}

\item{end_date}{character; calendar date in the format "yyyy-mm-dd"}
}
\value{
A list with components
\describe{
  \item{Cases_deaths}{A data frame at daily level, containing a date index, the date in format yyy-mm-dd",
                      the COVID-19 confirmed cumulative cases and cumulative deaths, the daily confirmed cases and the daily deaths.}
  \item{Date_range}{A vector of dates in the format "yyy-mm-dd", from the selected time series.}
}
}
\description{
Function to extract COVID-19 daily time series summary information, based on the
href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository} of the Johns
Hopkins University Center for Systems Science and Engineering (CSSE).
}
\examples{

\dontrun{
cd <- import_cases_deaths(country = "Greece",
                          start_date = "2020-03-05",
                          end_date   = "2020-10-07"
                          )
}

}
\references{
Dong, E., Du, H., & Gardner L. (2020). An interactive web-based dashboard to track COVID-19
in real time. Lancet Inf Dis, 20(5), 533-534. doi: 10.1016/S1473-3099(20)30120-1
}
