% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_infections.R
\name{plot_posterior_infections}
\alias{plot_posterior_infections}
\title{Plot the posterior distribution of the infection counts}
\usage{
plot_posterior_infections(
  object,
  type = c("age-specific", "aggregated"),
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{object}{A dataframe from \code{\link[Bernadette]{posterior_infections}}.}

\item{xlab}{character;
title of x-axis.}

\item{ylab}{character;
title of y-axis.}

\item{...}{Optional arguments passed to \code{\link[ggplot2]{facet_wrap}}, \code{\link[ggplot2]{scale_x_date}} and \code{\link[ggplot2]{theme}}.}
}
\value{
A \code{ggplot} object which can be further customised using the \pkg{ggplot2} package.
}
\description{
Plot the posterior distribution of the infection counts
}
\examples{
\dontrun{
if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
# Age-specific mortality/incidence count time series:
data(age_specific_mortality_counts)
data(age_specific_infection_counts)

# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece", year = 2020)

# Lookup table:
lookup_table <- data.frame(Initial = age_distr$AgeGrp,
                          Mapping = c("0-17",  "0-17",  "0-17",  "0-17",
                                      "18-39", "18-39", "18-39", "18-39",
                                      "40-64", "40-64", "40-64", "40-64", "40-64",
                                      "65+", "65+", "65+"))

# Aggregate the age distribution table:
aggr_age <- aggregate_age_distribution(age_distr, lookup_table)

# Aggregate the contact matrix:
aggr_cm <- aggregate_contact_matrix(cm,
                                    lookup_table,
                                    age.distr = aggr_age)

# Lookup table:
ifr_mapping <- c(rep("0-39", 8), rep("40-64", 5), rep("65+", 3))

# Aggregate the IFR:
aggr_age_ifr <- aggregate_ifr_react(age_distr, age_mapping, age_specific_infection_counts)

# Posterior sampling:
igbm_fit <- stan_igbm(y_data                      = age_specific_mortality_counts,
                      contact_matrix              = aggr_cm,
                      age_distribution_population = aggr_age,
                      age_specific_ifr            = aggr_age_ifr[[3]],
                      prior_volatility            = normal(location = 0, scale = 1),
                      prior_nb_dispersion         = gamma(shape = 2, rate = 1),
                      algorithm_inference         = "sampling")

post_inf_summary <- plot_posterior_infections(igbm_fit)

# Visualise the posterior distribution of the infection counts:
plot_posterior_infections(post_inf_summary, type = "age-specific")
plot_posterior_infections(post_inf_summary, type = "age-aggregated")
}
}
}
\seealso{
\code{\link{posterior_infections}}.
}
