% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt.R
\name{plot_posterior_rt}
\alias{plot_posterior_rt}
\title{Plot the estimated effective reproduction number trajectory
NOTE: See https://github.com/stan-dev/rstanarm/blob/master/R/posterior_traj.R}
\usage{
plot_posterior_rt(object, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{object}{A dataframe from \code{\link[Bernadette]{posterior_rt}}.}

\item{xlab}{character; Title of x-axis.}

\item{ylab}{character; Title of y-axis.}

\item{...}{Optional arguments passed to \code{\link[ggplot2]{scale_x_date}} and \code{\link[ggplot2]{theme}}.}
}
\value{
A \code{ggplot} object which can be further customised using the \pkg{ggplot2} package.
}
\description{
Plot the estimated effective reproduction number trajectory
NOTE: See https://github.com/stan-dev/rstanarm/blob/master/R/posterior_traj.R
}
\examples{
\dontrun{
if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
# Age-specific mortality/incidence count time series:
data(age_specific_mortality_counts)
data(age_specific_infection_counts)

# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece", year = 2020)

# Lookup table:
lookup_table <- data.frame(Initial = age_distr$AgeGrp,
                          Mapping = c(rep("0-39",  8),
                                      rep("40-64", 5),
                                      rep("65+"  , 3)))

# Aggregate the age distribution table:
aggr_age <- aggregate_age_distribution(age_distr, lookup_table)

# Import the projected contact matrix for a country (i.e. Greece):
conmat <- contact_matrix(country = "GRC")

# Aggregate the contact matrix:
aggr_cm <- aggregate_contact_matrix(conmat, lookup_table, aggr_age)

# Aggregate the IFR:
ifr_mapping <- c(rep("0-39", 8), rep("40-64", 5), rep("65+", 3))

aggr_age_ifr <- aggregate_ifr_react(age_distr, ifr_mapping, age_specific_infection_counts)

# Infection-to-death distribution:
ditd <- itd_distribution(ts_length  = nrow(age_specific_mortality_counts),
                         gamma_mean = 24.19231,
                         gamma_cv   = 0.3987261)

# Posterior sampling:
igbm_fit <- stan_igbm(y_data                      = age_specific_mortality_counts,
                      contact_matrix              = aggr_cm,
                      age_distribution_population = aggr_age,
                      age_specific_ifr            = aggr_age_ifr[[3]],
                      itd_distr                   = ditd,
                      likelihood_variance_type    = "quadratic",
                      prior_volatility            = normal(location = 0, scale = 1),
                      prior_nb_dispersion         = gamma(shape = 2, rate = 1),
                      algorithm_inference         = "sampling")

post_rt_summary <- posterior_rt(igbm_fit)

# Visualise the posterior distribution of the effective reproduction number:
plot_posterior_rt(post_rt_summary)
}
}
}
\seealso{
\code{\link{posterior_rt}}.
}
