% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeiir_mbm_ncp_gamma_volatilities.R
\name{seeiir_mbm_ncp_gamma_volatilities}
\alias{seeiir_mbm_ncp_gamma_volatilities}
\title{Multi-diffusion multi-type SEEIIR transmission model}
\usage{
seeiir_mbm_ncp_gamma_volatilities(standata, ...)
}
\arguments{
\item{standata}{A named `list` or `environment` providing the data for the model or a character vector for all the names of objects used as data. See \link[rstan]{sampling}.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains). See \link[rstan]{sampling}.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Function to implement the Bayesian multi-diffusion multi-type SEEIIR transmission model with Stan, assuming Gamma distributed prior distributions for the age-specific volatilities and a non-centered parameterisation of the latent diffusion path.
}
