% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeiir_sbm.R
\name{seeiir_sbm}
\alias{seeiir_sbm}
\title{Single-diffusion multi-type SEEIIR transmission model}
\usage{
seeiir_sbm(standata, ...)
}
\arguments{
\item{standata}{A named `list` or `environment` providing the data for the model or a character vector for all the names of objects used as data. See \link[rstan]{sampling}.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains). See \link[rstan]{sampling}.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Function to implement the Bayesian single-diffusion multi-type SEEIIR transmission model with Stan.
}
