% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stangbm.R
\name{summary.stanigbm}
\alias{summary.stanigbm}
\title{Summary of stanigbm posterior output}
\usage{
\method{summary}{stanigbm}(object, ...)
}
\arguments{
\item{object}{an \code{R} object of class \code{stanigbm}.}

\item{...}{Additional arguments, to be passed to lower-level functions. See \link[rstan]{summary-methods}.}
}
\description{
This function summarizes the MCMC output for \code{stanigbm} objects.
}
