% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_age_distribution.R
\name{aggregate_age_distribution}
\alias{aggregate_age_distribution}
\title{Aggregate the age distribution matrix}
\usage{
aggregate_age_distribution(x, lookup_table)
}
\arguments{
\item{x}{data.frame; an age distribution matrix. See \link[Bernadette]{age_distribution}.}

\item{lookup_table}{data.frame; a user-defined dataframe which maps the sixteen 5-year age bands to a new set of age bands.}
}
\value{
A data frame that contains the aggregated age distribution.
}
\description{
Function to aggregate the age distribution according to user-defined age groups.
}
\examples{

\dontrun{
# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece",
                             year    = 2020)

lookup_table <- data.frame(Initial = age_distr$AgeGrp,
                          Mapping = c("0-17",  "0-17",  "0-17",  "0-17",
                                      "18-39", "18-39", "18-39", "18-39",
                                      "40-64", "40-64", "40-64", "40-64", "40-64",
                                      "65+", "65+", "65+"))

# Aggregate the age distribution table:
aggr_age <- aggregate_age_distribution(age_distr,
                                      lookup_table)

# Plot the aggregated age distribution matrix:
plot_age_distribution(aggr_age)
}

}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2019). World Population Prospects 2019, Online Edition. Rev. 1.
}
