% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ifr_react.R
\name{aggregate_ifr_react}
\alias{aggregate_ifr_react}
\title{Aggregate the Infection Fatality Ratio}
\usage{
aggregate_ifr_react(x, user_AgeGrp)
}
\arguments{
\item{x}{data.frame; an age distribution matrix. See \link[Bernadette]{age_distribution}.}

\item{user_AgeGrp}{data.frame; a user-defined dataframe which maps the four age groups to a new set of age groups.}
}
\value{
A list of two data frames that contains the aggregated IFR estimates.
}
\description{
Function to aggregate the Infection Fatality Ratio (IFR) by age from REACT according to user-defined age groups.
}
\examples{

\dontrun{
# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece",
                             year    = 2020)

analysis_AgeGrp <- data.frame(REACT = c("15-44", "45-64", "65-74", "75+"),
                             User  = c(rep("0-44", 1),
                                      rep("45-64", 2),
                                      rep("65+",   1)))

# Aggregate the IFR:
aggr_age <- aggregate_ifr_react(age_distr, analysis_AgeGrp)

}

}
\references{
Ward, H., Atchison, C., Whitaker, M. et al. (2021). SARS-CoV-2 antibody prevalence in England following the first peak of the pandemic. Nature Communications 12, 905
}
