% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernadette-package.R
\docType{package}
\name{Bernadette-package}
\alias{Bernadette-package}
\alias{Bernadette}
\title{Bayesian inference and model selection for stochastic epidemics}
\description{
\code{Bernadette} provides Bayesian analysis for stochastic extensions of non-linear dynamic systems using advanced computational algorithms.
}
