% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernadette-package.R
\docType{package}
\name{Bernadette-package}
\alias{Bernadette-package}
\alias{Bernadette}
\title{Bayesian inference and model selection for stochastic epidemics}
\description{
\code{Bernadette} provides Bayesian analysis for stochastic extensions of non-linear dynamic systems using advanced computational algorithms.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R
package version 2.21.2. https://mc-stan.org
}
