% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_un.R
\name{countries_un}
\alias{countries_un}
\title{Names of countries with an available age distribution}
\usage{
countries_un()
}
\value{
A character vector that contains the names of 201 countries/areas.
}
\description{
Function to extract the names of the countries whose discrete age distribution is
available by the United Nations.
}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2019). World Population Prospects 2019, Online Edition. Rev. 1.

Prem, K., van Zandvoort, K., Klepac, P. et al (2017). Projecting contact matrices in 177 geographical regions: an update and comparison with empirical data for the COVID-19 era. medRxiv 2020.07.22.20159772; doi: https://doi.org/10.1101/2020.07.22.20159772
}
