% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contact_matrix.R
\name{plot_contact_matrix}
\alias{plot_contact_matrix}
\title{Contact matrix heatmap}
\usage{
plot_contact_matrix(x)
}
\arguments{
\item{x}{data.frame; a contact matrix.  See \link[Bernadette]{contact_matrix}.}
}
\value{
A heatmap of a contact matrix for a particular country.
}
\description{
Contact matrix heatmap
}
\examples{

\dontrun{

cm <- contact_matrix(country = "GRC")

plot_cm <- plot_contact_matrix(cm)
}

}
