% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_infections.R
\name{plot_posterior_infections}
\alias{plot_posterior_infections}
\title{Plot the posterior distribution of the infection counts}
\usage{
plot_posterior_infections(
  object,
  type = c("age-specific", "aggregated"),
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{object}{A dataframe from \code{\link[Bernadette]{posterior_infections}}.}

\item{type}{character;
Plot the output for the 'age-specific' infection counts or the 'aggregated' infections.}

\item{xlab}{character;
title of x-axis.}

\item{ylab}{character;
title of y-axis.}

\item{...}{Optional arguments passed to \code{\link[ggplot2]{scale_x_date}}.}
}
\value{
A \code{ggplot} object which can be further customised using the \pkg{ggplot2} package.
}
\description{
Plot the posterior distribution of the infection counts
}
\examples{
\donttest{
# Age-specific mortality/incidence count time series:
data(age_specific_mortality_counts)
data(age_specific_infection_counts)

# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece", year = 2020)

# Lookup table:
lookup_table <- data.frame(Initial = age_distr$AgeGrp,
                          Mapping = c(rep("0-39",  8),
                                      rep("40-64", 5),
                                      rep("65+"  , 3)))

# Aggregate the age distribution table:
aggr_age <- aggregate_age_distribution(age_distr, lookup_table)

# Import the projected contact matrix for a country (i.e. Greece):
conmat <- contact_matrix(country = "GRC")

# Aggregate the contact matrix:
aggr_cm <- aggregate_contact_matrix(conmat, lookup_table, aggr_age)

# Aggregate the IFR:
ifr_mapping <- c(rep("0-39", 8), rep("40-64", 5), rep("65+", 3))

aggr_age_ifr <- aggregate_ifr_react(age_distr, ifr_mapping, age_specific_infection_counts)

# Infection-to-death distribution:
ditd <- itd_distribution(ts_length  = nrow(age_specific_mortality_counts),
                         gamma_mean = 24.19231,
                         gamma_cv   = 0.3987261)

# Posterior sampling:

rstan_options(auto_write = TRUE)
chains <- 2
options(mc.cores = chains)

igbm_fit <- stan_igbm(y_data                      = age_specific_mortality_counts,
                      contact_matrix              = aggr_cm,
                      age_distribution_population = aggr_age,
                      age_specific_ifr            = aggr_age_ifr[[3]],
                      itd_distr                   = ditd,
                      incubation_period           = 3,
                      infectious_period           = 4,
                      likelihood_variance_type    = "linear",
                      ecr_changes                 = 7,
                      prior_scale_x0              = 0.5,
                      prior_scale_contactmatrix   = 0.05,
                      pi_perc                     = 0.1,
                      prior_volatility            = normal(location = 0, scale = 1),
                      prior_nb_dispersion         = exponential(rate = 1/5),
                      algorithm_inference         = "sampling",
                      nBurn                       = 5,
                      nPost                       = 10,
                      nThin                       = 1,
                      chains                      = chains,
                      adapt_delta                 = 0.8,
                      max_treedepth               = 16,
                      seed                        = 1)

post_inf_summary <- posterior_infections(object = igbm_fit,
                                         y_data = age_specific_mortality_counts)

# Visualise the posterior distribution of the infection counts:
plot_posterior_infections(post_inf_summary, type = "age-specific")
plot_posterior_infections(post_inf_summary, type = "aggregated")
}
}
\seealso{
\code{\link{posterior_infections}}.
}
