% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt.R
\name{posterior_rt}
\alias{posterior_rt}
\title{Estimate the effective reproduction number with the next generation matrix approach}
\usage{
posterior_rt(object)
}
\arguments{
\item{object}{An object of class \code{stanigbm}. See \code{\link[Bernadette]{stan_igbm}}.}
}
\value{
A ata.frames which can be visualised using \code{\link[Bernadette]{plot_posterior_rt}}.
}
\description{
Estimate the effective reproduction number with the next generation matrix approach
}
\examples{
\dontrun{
if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
# Age-specific mortality/incidence count time series:
data(age_specific_mortality_counts)
data(age_specific_infection_counts)

# Import the age distribution for a country in a given year:
age_distr <- age_distribution(country = "Greece", year = 2020)

# Lookup table:
lookup_table <- data.frame(Initial = age_distr$AgeGrp,
                          Mapping = c("0-17",  "0-17",  "0-17",  "0-17",
                                      "18-39", "18-39", "18-39", "18-39",
                                      "40-64", "40-64", "40-64", "40-64", "40-64",
                                      "65+", "65+", "65+"))

# Aggregate the age distribution table:
aggr_age <- aggregate_age_distribution(age_distr, lookup_table)

# Aggregate the contact matrix:
aggr_cm <- aggregate_contact_matrix(cm,
                                    lookup_table,
                                    age.distr = aggr_age)

# Lookup table:
ifr_mapping <- c(rep("0-39", 8), rep("40-64", 5), rep("65+", 3))

# Aggregate the IFR:
aggr_age_ifr <- aggregate_ifr_react(age_distr, age_mapping, age_specific_infection_counts)

# Posterior sampling:
igbm_fit <- stan_igbm(y_data                      = age_specific_mortality_counts,
                      contact_matrix              = aggr_cm,
                      age_distribution_population = aggr_age,
                      age_specific_ifr            = aggr_age_ifr[[3]],
                      prior_volatility            = normal(location = 0, scale = 1),
                      prior_nb_dispersion         = gamma(shape = 2, rate = 1),
                      algorithm_inference         = "sampling")

post_rt_summary <- posterior_rt(igbm_fit)

# Visualise the posterior distribution of the effective reproduction number:
plot_posterior_rt(post_rt_summary)
}
}
}
\references{
Diekmann, O., Heesterbeek, J., and Roberts, M. (2010). The construction of next-generation matrices for compartmental epidemic models. \emph{J. R. Soc. Interface}, 7, 873–-885.

Bouranis, L., Demiris, N. Kalogeropoulos, K. and Ntzoufras, I. (2022). Bayesian analysis of diffusion-driven multi-type epidemic models with application to COVID-19. arXiv: \url{https://arxiv.org/abs/2211.15229}
}
